♻️ Refactor. 이미지 처리 메서드를 익스텐션 파일로 분리
Why:
- BusDetectionViewController의 코드 길이 감소 및 가독성 향상
- 이미지 처리 로직을 별도 파일로 분리하여 관심사 분리
- 파일 구조 개선으로 유지보수성 향상

How:
- BusDetectionViewController+RefineImage.swift 익스텐션 파일 생성
- 이미지 크롭 및 리사이징 헬퍼 메서드를 새 파일로 이동
- 기능 변경 없이 코드 구조만 개선

Tag: #56
